1.re-fork the IBPSA repo,2 clean the repo such that only relevant contents are included,3.check if the convention is applied
