+++ IBPSA/Resources/Scripts/Dymola/Utilities/IO/RESTClient/Examples/OverWritten.mos
@@ -1,0 +1,7 @@
+// Script generated by Dymola Mon Jun 18 22:44:53 2018
+
+Advanced.CompileWith64=2
+simulateModel("IBPSA.Utilities.IO.RESTClient.Examples.OverWritten", stopTime=240, method="dassl", resultFile="OverWritten");
+createPlot(id=1, position={0, 0, 976, 426}, y={"sine1.y", "overWritten.y[1]"}, range={0.0, 240000.0, -3.0, 3.0}, grid=true, colors={{28,108,200}, {238,46,47}});
+createPlot(id=1, position={0, 0, 976, 210}, y={"sine2.y", "overWritten.y[2]"}, range={0.0, 240000.0, -1.5, 1.5}, grid=true, subPlot=2, colors={{28,108,200}, {238,46,47}});
+
