+++ IBPSA/Resources/Scripts/Dymola/Utilities/IO/RESTClient/Examples/Sampler.mos
@@ -1,0 +1,7 @@
+// Script generated by Dymola Mon Jun 18 22:44:53 2018
+
+Advanced.CompileWith64=2
+simulateModel("IBPSA.Utilities.IO.RESTClient.Examples.Sampler", stopTime=120, method="dassl", tolerance=1e-006, resultFile="Sampler");
+createPlot(id=1, position={0, 0, 976, 371}, y={"sine1.y", "sampler.oveSig[1]"}, range={0.0, 120000.0, -3.0, 3.0}, grid=true, colors={{28,108,200}, {238,46,47}});
+createPlot(id=1, position={0, 0, 976, 183}, y={"sine2.y", "sampler.oveSig[2]"}, range={0.0, 120000.0, -2.0, 2.0}, grid=true, subPlot=2, colors={{28,108,200}, {238,46,47}});
+
