+++ IBPSA/Utilities/IO/RESTClient/BaseClasses/package.order
@@ -1,1 +1,2 @@
-exchange
+SocClient
+PartialSocketClient
