+++ IBPSA/Utilities/IO/RESTClient/Examples/package.mo
@@ -1,0 +1,9 @@
+within IBPSA.Utilities.IO.RESTClient;
+package Examples "Collection of models that illustrate model use and test models"
+  extends Modelica.Icons.ExamplesPackage;
+
+
+annotation (Documentation(info="<html>
+<p>This package contains examples for the use of models that can be found in <code>IBPSA.Utilities.IO.RESTClient</code>. </p>
+</html>"));
+end Examples;
