+++ IBPSA/Utilities/IO/RESTClient/Types/package.mo
@@ -1,0 +1,9 @@
+within IBPSA.Utilities.IO.RESTClient;
+package Types "Package with type definitions"
+ extends Modelica.Icons.TypesPackage;
+
+
+annotation (preferredView="info", Documentation(info="<html>
+This package contains type definitions.
+</html>"));
+end Types;
