+++ IBPSA/Utilities/IO/RESTClient/Types/package.order
@@ -1,0 +1,2 @@
+GlobalActivation
+LocalActivation
