+++ IBPSA/Utilities/IO/RESTClient/package.order
@@ -1,2 +1,7 @@
-Overwrite_Real
+UsersGuide
+Types
+Examples
+Configuration
+OverWritten_Real
+Sampler_Real
 BaseClasses
